requirements/base.txt
Django>=4.1
djangorestframework
djangorestframework-simplejwt
django-environ
pillow